+++ android/app/src/main/java/org/mudamos/petition/firebase/Config.java
@@ -45,6 +45,7 @@ public class Config {
         defaults.put("link_help", "https://itsrio2.typeform.com/to/nGzwjv");
         defaults.put("link_send_your_idea", "https://itsrio2.typeform.com/to/iulNZI");
         defaults.put("link_why_projects", "https://www.mudamos.org/institucional/projetos-de-lei-de-iniciativa-popular");
+        defaults.put("link_privacy_policy", "https://www.mudamos.org/institucional/politica-de-privacidade");
         defaults.put("ineligible_to_sign_citywide_plip_reason", "Obrigado por seu apoio, mas você só pode assinar esse projeto de lei sendo eleitor do município para o qual ele se destina. Se deseja propor essa lei para o seu município, use a função \"Proponha um PL\" no menu do App.");
         defaults.put("ineligible_to_sign_statewide_plip_reason", "Obrigado por seu apoio, mas você só pode assinar esse projeto de lei sendo eleitor do estado para o qual ele se destina. Se deseja propor essa lei para o seu município, use a função \"Proponha um PL\" no menu do App.");